+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Add support for Zergpool
 * Add support for Bitcore, BitSend, Creativecoin, Folm Coin, Hshare, LUXCoin, Solaris, Trezarcoin
 * Add support for TimeTravel10, Xevan
-* Add support for ccminer-xevan
+* Add support for ccminer-xevan, hsrminer-neoscrypt
 * Don't require a restart when changing the region
 * Re-download miner if deleted during mining
 * Improve First Run Wizard